Removing aditional files and leaving all in ther Mastermind.cs as requested
